% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_fanchart}
\alias{add_fanchart}
\title{Add fanchart to ggplot}
\usage{
add_fanchart(
  df,
  x = "Time",
  ymin = "Lower",
  ymax = "Upper",
  fill = "Level",
  legend_fill = c("continuous", "discrete"),
  labs_fill = ifelse(legend_fill == "continuous", "Confidence level", "Probability"),
  palette = c("#EFF3FF", "#C6DBEF", "#9ECAE1", "#6BAED6", "#3182BD", "#08519C")
)
}
\arguments{
\item{df}{Data with columns \code{x}, \code{ymin}, \code{ymax} and \code{fill}}

\item{x}{Name of the \code{x} aesthetic}

\item{ymin}{Name of the \code{ymin} aesthetic}

\item{ymax}{Name of the \code{ymax} aesthetic}

\item{fill}{Name of the \code{fill} aesthetic}

\item{legend_fill}{Whether the legend should be displayed as \code{continuous} or as \code{discrete} categories}

\item{labs_fill}{Name to give to the legend}

\item{palette}{Colour palette to use.
The default is the single-hue blue palette from \code{RColorBrewer::brewer.pal(n = 6, "Blues")}.}
}
\value{
List to be added to a ggplot
}
\description{
The fanchart is obtained by overlaying \code{\link[ggplot2:geom_ribbon]{ggplot2::geom_ribbon()}} of different widths (corresponding to different levels).
NB: this function is not a geom.
}
\section{Alternative}{

A similar result can be obtained using the \code{\link[ggdist:geom_lineribbon]{ggdist::geom_lineribbon()}} with the difference that the \code{ggdist} function also plots a point estimate (and is a proper geom).
}

\examples{

library(dplyr)
library(tidyr)
library(ggplot2)

tmp <- tibble(Time = 0:10,
              y = Time^1.5) \%>\%
  expand_grid(Level = seq(0.1, 0.9, 0.2)) \%>\%
  mutate(Width = qnorm(0.5 + Level / 2, sd = 2),
         Lower = y - Width,
         Upper = y + Width)
ggplot() + add_fanchart(tmp)

}
