% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{add_historical_pred}
\alias{add_historical_pred}
\title{Performance of (population) historical forecast}
\usage{
add_historical_pred(
  test,
  train,
  max_score,
  discrete = TRUE,
  add_uniform = TRUE,
  include_samples = FALSE,
  n_samples = NULL
)
}
\arguments{
\item{test}{Testing dataframe. The only requirements is that it contains a column "Score".}

\item{train}{Training dataframe. The only requirements is that it contains a column "Score".}

\item{max_score}{Maximum value that the score can take}

\item{discrete}{Whether to estimate a discrete or continuous historical forecast}

\item{add_uniform}{Whether to include samples from uniform distribution when computing a discrete historical forecast.
This ensures that all states are visited.}

\item{include_samples}{Whether to return samples from the historical forecast in the output}

\item{n_samples}{If \code{include_samples=TRUE}, how many samples to return. When NULL, the function return the training set.}
}
\value{
Dataframe \code{test} appended by the columns "lpd", "RPS" (or CRPS if discrete=FALSE) and optionally "Samples"
}
\description{
Performance of (population) historical forecast
}
\details{
The continuous historical forecast is calculated by considering the training set as samples from the predictive distribution.
}
\examples{
max_score <- 100
train <- data.frame(Score = rbinom(1e2, max_score, 0.2))
test <- data.frame(Score = rbinom(1e2, max_score, 0.5))
add_historical_pred(test, train, max_score)
}
