% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{add_predictions}
\alias{add_predictions}
\title{Append lpd, (C)RPS and predictive samples to testing dataframe}
\usage{
add_predictions(
  test,
  fit,
  discrete = TRUE,
  include_samples = FALSE,
  n_samples = NULL
)
}
\arguments{
\item{test}{Testing dataframe.
The only requirements is that it contains a column "Score" when discrete=FALSE.}

\item{fit}{Stanfit object}

\item{discrete}{Whether to estimate a discrete or continuous forecast.
For a discrete forecast, the RPS will be computed and the CRPS for a continuous forecast.}

\item{include_samples}{Whether to return samples from the historical forecast in the output}

\item{n_samples}{If include_samples=TRUE, how many samples to return. Default (=NULL) to all samples.}
}
\value{
Dataframe test appended by the columns "lpd", "RPS" (or CRPS if discrete=FALSE) and optionally "Samples"
}
\description{
Append lpd, (C)RPS and predictive samples to testing dataframe
}
