% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{add_uniform_pred}
\alias{add_uniform_pred}
\title{Performance of a uniform forecast}
\usage{
add_uniform_pred(
  test,
  max_score,
  discrete = TRUE,
  include_samples = FALSE,
  n_samples = NULL
)
}
\arguments{
\item{test}{Testing dataframe. The only requirements is that it contains a column "Score".}

\item{max_score}{Maximum value that the score can take}

\item{discrete}{Whether to estimate a discrete or continuous historical forecast}

\item{include_samples}{Whether to return samples from the historical forecast in the output}

\item{n_samples}{If include_samples=TRUE, how many samples to return.
When NULL, the samples \code{0:max_score} are returned.}
}
\value{
Dataframe \code{test} appended by the columns "lpd", "RPS" (or CRPS if discrete=FALSE) and optionally "Samples"
}
\description{
Performance of a uniform forecast
}
\examples{
max_score <- 100
test <- data.frame(Score = rbinom(1e2, max_score, 0.5))
add_uniform_pred(test, max_score)
}
