% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR1.R
\name{AR1}
\alias{AR1}
\alias{default_prior_AR1}
\alias{fit_AR1}
\alias{sample_prior_AR1}
\title{Autoregressive model (order 1)}
\usage{
default_prior_AR1()

fit_AR1(train, test = NULL, max_score, prior = default_prior_AR1(), ...)

sample_prior_AR1(
  N_patient = 1,
  t_max = c(2),
  max_score,
  prior = default_prior_AR1(),
  ...
)
}
\arguments{
\item{train}{Training dataframe (details of the format in \code{\link[=prepare_data_lgtd]{prepare_data_lgtd()}})}

\item{test}{Testing dataframe (details of the format in \code{\link[=prepare_data_lgtd]{prepare_data_lgtd()}})}

\item{max_score}{Maximum value that the score can take}

\item{prior}{A named list with elements \code{sigma}, \code{y_inf}, \code{alpha} specifying priors
for the corresponding parameters, such as:
\itemize{
\item \code{sigma / max_score ~ normal(x1, x2)}.
The element \code{sigma} of the list should be a vector of length 2 containing x1 and x2.
NB: usually x1=0 to define a half-normal distribution (sigma is constraint to be positive) and
x2 should be positive.
\item \code{y_inf / max_score ~ N(x1, x2)}.
The element \code{y_inf} of the list should be a vector of length 2 containing x1 and x2, x2 should be positive.
\item \code{alpha ~ beta(x1, x2)}.
The element \code{alpha} of the list should be a vector of length 2 containing x1 and x2, both positives.
}}

\item{...}{Arguments to be pass to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}

\item{N_patient}{Number of patients}

\item{t_max}{Vector of size N_patient indicating the time-series length}
}
\value{
\itemize{
\item \code{default_prior_AR1}: named list defining the default priors (see details)
\item \code{fit_AR1}: Stanfit object
\item \code{sample_prior_AR1}: Stanfit object
}
}
\description{
\itemize{
\item \code{default_prior_AR1} returns the default prior
\item \code{fit_AR1} fits the AR1 model
\item \code{sample_prior_AR1} samples the prior predictive distribution
}
}
\details{
\itemize{
\item The default prior for \code{sigma} translates to a width of the predictive distribution to be at most \code{max_score}.
\item The default prior for \code{y_inf} covers the full range of the score.
\item The default prior for \code{alpha} is uniform in 0-1.
\item The model is naive as it is trained with a non-truncated distribution
\item For more details see the \url{vignette}.
}
}
\section{Model's parameters}{

\itemize{
\item \code{sigma}: Standard deviation of the autoregression
\item \code{alpha}: Autocorrelation parameter
\item \code{b}: Intercept
\item \code{y_inf}: Autoregression mean
\item \code{y_mis}: Missing values
}

See \code{list_parameters(model = "AR1")}.
}

