% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_doc.R
\name{BinMRW}
\alias{BinMRW}
\title{Binomial Multivariate Random Walk model}
\arguments{
\item{max_score}{Maximum value that the scores can take}

\item{D}{Number of components}

\item{independent_components}{Whether to have diagonal correlations matrices or not}

\item{prior}{Named list of the model's priors.
It uses the default priors (see \code{\link[=default_prior]{default_prior()}}) if \code{NULL} and for the parameters that are not provided.}
}
\description{
This is a state-space model defined by a Binomial measurement error and a latent multivariate random walk.
}
\details{
Details of the model are available in the \href{#}{paper}.
}
\section{Parameters}{


The parameters are essentially the same as the \link{BinRW} model,
except that they one extra dimension corresponding to the component.
In addition, there are two news population parameters:
\itemize{
\item \code{Omega}: Correlation matrix of the latent dynamic
\item \code{Sigma}: Covariance matrix of the latent dynamic
\item \code{Omega0}: Correlation matrix of the initial condition (cf. \code{logit_y0})
}

And one extra observation (patient- and time-dependent) parameter:
\itemize{
\item \code{ytot_rep}, which corresponds to the posterior replications of the sum of all components
}

See \code{list_parameters(model = "BinMRW")} for more details.
}

\section{Priors}{

The priors are passed as a named list with elements \code{Omega}, \code{Omega0}, \code{sigma}, \code{mu_logit_y0} and \code{sigma_logit_y0}
specifying priors for the corresponding parameters.

\code{Omega} and \code{Omega0} follow a LKJ distribution with shape parameter \code{eta} > 0.
The shape parameter \code{eta} can be interpreted like the shape parameter of a symmetric beta distribution.
When \code{eta} is 1, this corresponds to a "uniform" distribution over correlation matrices.
\verb{eta >> 1} converges to identity matrix.
\itemize{
\item \code{Omega ~ LJJ(eta)}
\item \code{Omega0 ~ LKJ(eta)}
}

The elements \code{sigma}, \code{mu_logit_y0} and \code{sigma_logit_y0} should be a matrix with 2 rows and \code{D} columns,
where the d-th column {x1 x2} correspond to the prior for the d-th component, with x2 > 0 and:
\itemize{
\item \verb{sigma[d] ~ normal+(x1, x2)}.
\item \code{mu_logit_y0[d] ~ normal(x1, x2)}
\item \verb{sigma_logit_y0[d] ~ normal+(x1, x2)}
}

NB: For \code{sigma} and \code{sigma_logit_y0}, usually x1=0 to define a half-normal distribution
since the parameters are constrained to be positive.
}

\section{Default priors}{

\itemize{
\item The default priors for each \code{sigma}, \code{mu_logit_y0} and \code{sigma_logit_y0}
are the same as the default priors for \link{BinRW}.
\item The default priors for \code{Omega} and \code{Omega0} is a LKJ distribution with a shape parameter of 1.
}
}

\examples{
EczemaMVModel("BinMRW", max_score = 10, D = 5)
}
