% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinRW.R
\name{BinRW}
\alias{BinRW}
\alias{default_prior_BinRW}
\alias{fit_BinRW}
\alias{sample_prior_BinRW}
\title{Binomial random walk model}
\usage{
default_prior_BinRW()

fit_BinRW(train, test = NULL, max_score, prior = default_prior_BinRW(), ...)

sample_prior_BinRW(
  N_patient = 1,
  t_max = c(2),
  max_score,
  prior = default_prior_BinRW(),
  ...
)
}
\arguments{
\item{train}{Training dataframe (details of the format in \code{\link[=prepare_data_lgtd]{prepare_data_lgtd()}})}

\item{test}{Testing dataframe (details of the format in \code{\link[=prepare_data_lgtd]{prepare_data_lgtd()}})}

\item{max_score}{Maximum value that the score can take}

\item{prior}{A named list with elements \code{sigma}, \code{mu_logit_y0} and \code{sigma_logit_y0}
specifying priors for the corresponding parameters.
Each element of the list should be a vector of length 2, containing values for x1 and x2, x2 > 0, such as:
\itemize{
\item \code{sigma ~ normal(x1, x2)}.
\item \code{mu_logit_y0 ~ normal(x1, x2)}
\item \code{sigma_logit_y0 ~ normal(x1, x2)}
}

NB: For \code{sigma} and \code{sigma_logit_y0}, usually x1=0 to define a half-normal distribution
since the parameters are constrained to be positive.}

\item{...}{Arguments to be pass to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}

\item{N_patient}{Number of patients}

\item{t_max}{Vector of size N_patient indicating the time-series length}
}
\value{
\itemize{
\item \code{default_prior_BinRW}: named list defining the default priors
\item \code{fit_BinRW}: Stanfit object
\item \code{sample_prior_BinRW}: Stanfit object
}
}
\description{
This is a state-space model defined by a Binomial measurement error and a latent random walk.
}
\details{
\itemize{
\item \code{default_prior_BinRW} returns the default prior
\item \code{fit_BinRW} fits the Binomial random walk model
\item \code{sample_prior_BinRW} samples the prior predictive distribution
\item The default prior for \code{sigma} translates to an odd ratio increment of at most 5 (~ 2 * upper bound of prior).
\item The default priors for \code{mu_logit_y0} and \code{sigma_logit_y0} translates to an approximately uniform prior on \code{y0}.
\item For more details see the \url{vignette}.
}
}
\section{Model's parameters}{


Population parameters:
\itemize{
\item \code{sigma}: Standard deviation of the random walk (in logit scale)
\item \code{mu_logit_y0}: Population mean of the initial condition (in logit scale)
\item \code{sigma_logit_y0}: Population standard deviation of the initial condition (logit scale)
}

Patient-dependent parameters:
\itemize{
\item \code{logit_y0}: Logit of the initial condition
}

Observation-dependent (patient- and time-dependent) parameters:
\itemize{
\item \code{y_lat}: Latent score (probability)
\item \code{logit_lat}: logit of \code{y_lat}
}

See \code{list_parameters(model = "BinRW")}.
}

