% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prior.R
\name{default_prior.BinMRW}
\alias{default_prior.BinMRW}
\title{Default prior for BinMRW}
\usage{
\method{default_prior}{BinMRW}(model, ...)
}
\arguments{
\item{model}{Object of class "BinMRW"}

\item{...}{Arguments to pass to other methods}
}
\value{
List of parameters priors
}
\description{
Default prior for BinMRW
}
