% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{default_prior}
\alias{default_prior}
\title{Call default_prior_* for different models}
\usage{
default_prior(
  model = c("BinRW", "BinMC", "OrderedRW", "RW", "Smoothing", "AR1", "MixedAR1"),
  max_score = NULL
)
}
\arguments{
\item{model}{Model name}

\item{max_score}{Maximum value that the score can take (only required for model = "OrderedRW")}
}
\value{
List (see default_prior_*)
}
\description{
Call default_prior_* for different models
}
\examples{
default_prior(model = "BinRW")
}
