% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R
\name{EczemaModel}
\alias{EczemaModel}
\title{EczemaModel constructor}
\usage{
EczemaModel(
  model_name = c("BinRW"),
  max_score,
  discrete = (model_name \%in\% c("BinRW")),
  K = ifelse(discrete, max_score + 1, NA),
  prior = NULL
)
}
\arguments{
\item{model_name}{Name of the model to create}

\item{max_score}{Maximum value that the score can take}

\item{discrete}{Whether the model is discrete or not}

\item{K}{Number of categories (only used when \code{model_name == "MC"})}

\item{prior}{Prior for the model. If \code{NULL}, uses the default prior for the model (see \code{\link[=default_prior]{default_prior()}})}
}
\value{
An object (list) of class \code{model_name} and EczemaModel
}
\description{
EczemaModel constructor
}
\examples{
EczemaModel("BinRW", max_score = 10)
}
