% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MV_model_class.R
\name{EczemaMVModel}
\alias{EczemaMVModel}
\title{EczemaMVModel constructor}
\usage{
EczemaMVModel(
  model_name = c("BinMRW", "OrderedMRW"),
  max_score,
  D,
  independent_components = FALSE,
  prior = NULL
)
}
\arguments{
\item{model_name}{Name of the model to create}

\item{max_score}{Maximum value that the scores can take}

\item{D}{Number of components}

\item{independent_components}{Whether to have diagonal correlations matrices or not.
In other words, whether changes in latent scores are correlated,
or whether the components are indepedent.}

\item{prior}{Named list of the model's priors.
It uses the default priors (see \code{\link[=default_prior]{default_prior()}}) if \code{NULL} and for the parameters that are not provided.}
}
\value{
An object of class \code{model_name} and EczemaMVModel
}
\description{
EczemaMVModel constructor
}
\examples{
EczemaMVModel("BinMRW", max_score = 10, D = 5)
}
