% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_2d_parname}
\alias{expand_2d_parname}
\title{Full names of a two-dimensional parameter}
\usage{
expand_2d_parname(par_name, n1, n2)
}
\arguments{
\item{par_name}{Name of the parameter}

\item{n1}{First dimension}

\item{n2}{Second dimension}
}
\value{
Character vector
}
\description{
Full names of a two-dimensional parameter
}
