% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{extract_metric}
\alias{extract_metric}
\alias{extract_lpd}
\alias{extract_RPS}
\title{Extract lpd and RPS from stanfit object}
\usage{
extract_lpd(fit)

extract_RPS(fit)
}
\arguments{
\item{fit}{Stanfit object}
}
\value{
Vector of lpd/RPS for each prediction
}
\description{
The lpd and RPS are computed for the expected forecast distribution.
The lpd is defined for continuous and discrete outcomes.
The RPS is defined for discrete outcomes only and is computed by
extracting the cumulative error distribution (cumulative forecast - cumulative distribution),
for which we take the expected value, square it and apply and \code{sum() / (R-1)}.
}
