% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{extract_parameters}
\alias{extract_parameters}
\title{Extract parameters posterior summary statistics}
\usage{
extract_parameters(fit, pars = NULL, id = NULL, ...)
}
\arguments{
\item{fit}{Stanfit object}

\item{pars}{Named list of parameters to extract. See \code{\link[=list_parameters]{list_parameters()}}.}

\item{id}{Dataframe associating Index to (Patient, Time) pairs. See \code{\link[=get_index]{get_index()}}.}

\item{...}{Arguments to pass to \code{\link[HuraultMisc:summary_statistics]{HuraultMisc::summary_statistics()}}.}
}
\value{
Tibble dataframe containing posterior summary statistics.
See details in \code{\link[HuraultMisc:summary_statistics]{HuraultMisc::summary_statistics()}}. Additional columns Patient and Time if \code{id} is not NULL.
}
\description{
Extract parameters posterior summary statistics
}
