% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_longitudinal.R
\name{extract_simulations}
\alias{extract_simulations}
\title{Extract simulations}
\usage{
extract_simulations(fit, id, draw, pars = NULL)
}
\arguments{
\item{fit}{Stanfit object}

\item{id}{Dataframe linking index in Stan model to (Patient, Time) pairs, cf. output from \code{\link[=get_index]{get_index()}}}

\item{draw}{Draw ID}

\item{pars}{Vector of parameters to extract. Default to all parameters except \code{y_rep}.}
}
\value{
Named list:
\itemize{
\item Data: dataframe with columns Patient, Time, Index, and Score corresponding to simulations
\item Parameters: dataframe containing parameters used to generate the data (cf. \code{\link[HuraultMisc:extract_draws]{HuraultMisc::extract_draws()}})
}
}
\description{
Extract simulations
}
\details{
This function is not desgined to use with the Markov Chain model (MC).
}
