% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{fit_continuous}
\alias{fit_continuous}
\title{Fit continuous models}
\usage{
fit_continuous(
  train,
  test = NULL,
  max_score,
  model = c("RW", "Smoothing", "AR1", "MixedAR1"),
  prior = default_prior(model, max_score),
  ...
)
}
\arguments{
\item{train}{Training dataframe (details of the format in \code{\link[=prepare_data_lgtd]{prepare_data_lgtd()}})}

\item{test}{Testing dataframe (details of the format in \code{\link[=prepare_data_lgtd]{prepare_data_lgtd()}})}

\item{max_score}{Maximum value that the score can take}

\item{model}{Model name}

\item{prior}{A named list of corresponding to parameters' prior.}

\item{...}{Arguments to be pass to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}
}
\value{
Stanfit object
}
\description{
Fit continuous models
}
\seealso{
\code{\link[=fit_RW]{fit_RW()}}, \code{\link[=fit_Smoothing]{fit_Smoothing()}}, \code{\link[=fit_AR1]{fit_AR1()}}, \code{\link[=fit_MixedAR1]{fit_MixedAR1()}}
}
