% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{fit_discrete}
\alias{fit_discrete}
\title{Fit discrete models}
\usage{
fit_discrete(
  train,
  test = NULL,
  max_score,
  model = c("BinRW", "BinMC", "OrderedRW", "RW"),
  prior = default_prior(model, max_score),
  ...
)
}
\arguments{
\item{train}{Training dataframe (details of the format in \code{\link[=prepare_data_lgtd]{prepare_data_lgtd()}})}

\item{test}{Testing dataframe (details of the format in \code{\link[=prepare_data_lgtd]{prepare_data_lgtd()}})}

\item{max_score}{Maximum value that the score can take}

\item{model}{Model name}

\item{prior}{A named list of corresponding to parameters' prior.}

\item{...}{Arguments to be pass to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}
}
\value{
Stanfit object
}
\description{
Except Markov Chain model
}
\seealso{
\code{\link[=fit_BinRW]{fit_BinRW()}}, \code{\link[=fit_BinMC]{fit_BinMC()}}, \code{\link[=fit_OrderedRW]{fit_OrderedRW()}}, \code{\link[=fit_RW]{fit_RW()}}
}
