% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{forward_chaining}
\alias{forward_chaining}
\alias{get_fc_iteration}
\alias{split_fc_dataset}
\alias{detail_fc_training}
\alias{get_fc_training_iteration}
\title{Forward chaining}
\usage{
get_fc_iteration(t, horizon)

split_fc_dataset(df, it)

detail_fc_training(df, horizon)

get_fc_training_iteration(it_test)
}
\arguments{
\item{t}{Vector of timepoints}

\item{horizon}{Updating horizon}

\item{df}{Dataframe with columns Patient, Time, Horizon, Iteration.
Only the colums Time is required for \code{detail_fc_training}.}

\item{it}{Iteration number}

\item{it_test}{Vector of testing iteration numbers}
}
\value{
\itemize{
\item \code{get_fc_iteration} returns a vector corresponding to iteration numbers
\item \code{split_fc_dataset} returns a named list of dataframes ("Training" and "Testing")
\item \code{get_fc_training_iteration} returns a vector of unique training iteration number
\item \code{detail_fc_training} returns a dataframe with columns: Iteration, N, Proportion, LastTime
}
}
\description{
In a forward chaining setting, we retrain/update the model every \code{horizon} day/week/months/etc.
and test the model on the next \code{horizon} day/week/months/etc., i.e. the data that is included at the next training step.
We call this training/testing steps iterations and start at iteration 0 where only the first timepoint is only included.
At iteration 1, timepoints 1 to \code{1 + horizon} are included.
}
\details{
\itemize{
\item \code{get_fc_iteration} associate a vector of timepoints \code{t} to the corresponding iteration.
\item \code{split_fc_dataset} split dataset \code{df} into a training and testing set,
and computes prediction horizon and last available score in the test set.
\item \code{get_fc_training_iteration} identify training iterations such as test set is not empty
\item \code{detail_fc_training} derive training data characteristics for each iteration in \code{df},
including the number of training observations, the proportion of training observations on the total number of observations,
and the maximum timepoint in the training sets.
}

Time=t is in Iteration=i means that:
\itemize{
\item Time=t is in the new training data of iteration i
\item Time=t is not in the training data of iterations < i
\item Time=t is in the testing data of iteration i-1
\item Time=t is in the training data of iterations >= i
}
}
\examples{
h <- 2
df <- get_index2(N_patient = 10, t_max = rpois(10, 10))
df$Score <- rnorm(nrow(df))
df$Iteration <- get_fc_iteration(df$Time, h)
sp <- split_fc_dataset(df, 1)
train_it <- get_fc_training_iteration(df$Iteration)
fc_char <- detail_fc_training(df, h)
}
