% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{generate_missing}
\alias{generate_missing}
\title{Generate missing values in a times-series}
\usage{
generate_missing(
  N,
  type = c("random", "markovchain"),
  p_mis = 0.25,
  p_obs_obs = 0.75
)
}
\arguments{
\item{N}{Length of the time-series}

\item{type}{Method to generate the missing values. One of "random" or "markovchain"}

\item{p_mis}{Probability of a given value to be missing (steady state probability for type == "markovchain")}

\item{p_obs_obs}{Probability of the next value being observed when the current is observed (for type == "markovchain")}
}
\value{
Logical vector of length N
}
\description{
First and last values are nor missing.
Missing indices can be generated at random (Binomial distribution) or using a Markov Chain (if consecutives missing values are deemed more likely).
The markov chain is parametrised in terms of the steady state probability of a value being missing and the probability that the next value is observed when the current value is also observed.
}
\examples{
generate_missing(10)
generate_missing(10, type = "markovchain")
}
