% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_compiled_model}
\alias{get_compiled_model}
\title{Get compiled model}
\usage{
get_compiled_model(stanmodel)
}
\arguments{
\item{stanmodel}{Stan model name.
NB: this may differ from the name of the model}
}
\value{
Compiled model (object to pass to \code{rstan::sampling})
}
\description{
Get compiled model
}
