% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal.R
\name{get_index}
\alias{get_index}
\alias{get_index2}
\title{Associate (Patient, Time) pairs to corresponding index in the model}
\usage{
get_index(train, test = NULL)

get_index2(N_patient, t_max)
}
\arguments{
\item{train}{Training dataframe}

\item{test}{Testing dataframe}

\item{N_patient}{Number of patients}

\item{t_max}{Vector of size N_patient indicating the time-series length}
}
\value{
Dataframe with columns Patient, Time, Index
}
\description{
Associate (Patient, Time) pairs to corresponding index in the model
}
\details{
For more details, see the \url{vignette}
}
\examples{
library(dplyr)
id <- get_index2(N_patient = 10, t_max = rpois(10, 20))
df <- id \%>\% select(-Index) \%>\% slice_sample(prop = 0.9) \%>\% arrange(Patient, Time)
get_index(train = df)
}
