% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{list_parameters}
\alias{list_parameters}
\title{List available parameters}
\usage{
list_parameters(
  model = c("MC", "BinRW", "BinMC", "OrderedRW", "RW", "Smoothing", "AR1", "MixedAR1"),
  main = TRUE
)
}
\arguments{
\item{model}{Model name}

\item{main}{Whether to output the main parameters only.
Parameters that are simple transformation of another parameter are dropped.
For instance, in BinRW, \code{logit_lat} would be dropped but \code{y_lat} would remain.}
}
\value{
Named list of parameters names, grouped into broad categories:
\itemize{
\item Population: population parameters (i.e. patient- and time-independent)
\item Patient: patient-dependent parameters
\item PatientTime: patient- and time-dependent parameters (e.g. latent scores)
\item Test: parameters related to the test set
\item Misc: other parameters
}
}
\description{
List available parameters
}
\details{
See \link{MC}, \link{BinRW}, \link{BinMC}, \link{OrderedRW}, \link{RW}, \link{Smoothing}, \link{AR1} and \link{MixedAR1} for details about the model-specific parameters.
Other parameters are available across models:
\itemize{
\item \code{y_rep} correspond to posterior replications. To get the corresponding index, use \code{\link[=get_index]{get_index()}}.
\item \code{y_pred} is a subset of y_rep corresponding to test samples (size \code{N_test} equal to the number of observations in the test set).
\item \code{lpd} is the log predictive density of test samples (of size \code{N_test}).
\item \code{cum_err} is the cumulative error distribution, only available for discrete outcomes
(matrix with dimensions \code{N_test * (max_score  + 1)}).
}
}
\examples{
list_parameters("BinRW")
}
