% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_inference.R
\name{make_empty_data}
\alias{make_empty_data}
\title{Create "empty" train and test set to pass to \code{prepare_standata} for \code{sample_prior}}
\usage{
make_empty_data(N_patient = 1, t_max = c(2), max_score, discrete)
}
\arguments{
\item{N_patient}{Number of patients in the datasets}

\item{t_max}{Vector of size \code{N_patient} indicating the time-series length of each patient}

\item{max_score}{Maximum value that the score can take}

\item{discrete}{Whether to use a discrete normal distribution (only relevant for testing)}
}
\value{
List containing "Training" and "Testing" dataframes
}
\description{
Used internally.
}
\examples{
make_empty_data(max_score = 10, discrete = TRUE)
}
