% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC.R, R/model_doc.R
\name{prepare_data_MC}
\alias{prepare_data_MC}
\alias{MC}
\title{Markov Chain model}
\usage{
prepare_data_MC(train, test = NULL, K)
}
\arguments{
\item{K}{Number of states of the Markov Chain}

\item{prior}{Named list of the model's priors. If \code{NULL}, uses the default prior for the model (see \code{\link[=default_prior]{default_prior()}}).}
}
\description{
For more details see the \href{https://ghurault.github.io/EczemaPred/articles/MC.html}{Markov Chain vignette}.
}
\details{
\itemize{
\item \code{train} and \code{test} should have columns \code{y0} (for the current state), \code{y1} (for the next state) and
\code{dt} (for the time delay between states).
\item \code{y0} and \code{y1} should take integer values between 1 and K.
\item \code{dt} should take integer values greater than or equal to 1.
}

Missing values are not allowed.

Details of the model are available in the \href{#}{paper}.
}
\section{Parameters}{

\itemize{
\item \code{p}: matrix of size K * K where \code{p[i, j]} represents the transition probabilities from state i to state j.
}

See \code{list_parameters(model = "MC")}.
}

\section{Priors}{

The priors are passed as a named list with element \code{p}.
The transition probabilities from state i \code{p[i, ]} are assumed to follow a Dirichlet distribution.
The prior should be a matrix where each line correspond to the parameters of the Dirichlet distribution for \code{p[i, ]}.
}

\section{Default priors}{

The default prior for all \code{p[i, ]} is a symmetric uniform Dirichlet distribution (all concentration parameters are equal to 1).
}

\examples{
EczemaModel("MC", K = 5)
}
