% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedAR1.R
\name{MixedAR1}
\alias{MixedAR1}
\alias{default_prior_MixedAR1}
\alias{fit_MixedAR1}
\alias{sample_prior_MixedAR1}
\title{Mixed effect autoregressive model (order 1)}
\usage{
default_prior_MixedAR1()

fit_MixedAR1(
  train,
  test = NULL,
  max_score,
  prior = default_prior_MixedAR1(),
  ...
)

sample_prior_MixedAR1(
  N_patient = 1,
  t_max = c(2),
  max_score,
  prior = default_prior_MixedAR1(),
  ...
)
}
\arguments{
\item{train}{Training dataframe (details of the format in \code{\link[=prepare_data_lgtd]{prepare_data_lgtd()}})}

\item{test}{Testing dataframe (details of the format in \code{\link[=prepare_data_lgtd]{prepare_data_lgtd()}})}

\item{max_score}{Maximum value that the score can take}

\item{prior}{A named list with elements \code{sigma}, \code{mu_logit_alpha}, \code{sigma_logit_alpha}, \code{mu_inf}, \code{sigma_inf}
specifying priors for the corresponding parameters.
Each element of the list should be a vector of length 2, containing values for x1 and x2, x2 > 0, such as:
\itemize{
\item \code{sigma / max_score ~ normal(x1, x2)}.
\item \code{mu_logit_alpha ~ normal(x1, x2)}.
\item \code{sigma_logit_alpha ~ normal(x1, x2)}.
\item \code{mu_inf / max_score ~ normal(x1, x2)}.
\item \code{sigma_inf / max_score ~ normal(x1, x2)}.
}

NB: For \code{sigma}, \code{sigma_logit_alpha} and \code{sigma_inf}, usually x1=0 to define a half-normal distribution
since the parameter is constrained to be positive.}

\item{...}{Arguments to be pass to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}

\item{N_patient}{Number of patients}

\item{t_max}{Vector of size N_patient indicating the time-series length}
}
\value{
\itemize{
\item \code{default_prior_MixedAR1}: named list defining the default priors (see details).
\item \code{fit_MixedAR1}: Stanfit object
\item \code{sample_prior_MixedAR1}: Stanfit object
}
}
\description{
\itemize{
\item \code{default_prior_MixedAR1} returns the default prior
\item \code{fit_MixedAR1} fits the mixed AR1 model
\item \code{sample_prior_MixedAR1} samples the prior predictive distribution
}
}
\details{
\itemize{
\item The default prior for \code{sigma} translates to a width of the predictive distribution to be at most \code{max_score}.
\item The default priors for \code{mu_logit_alpha} and \code{sigma_logit_alpha} have "reasonable" ranges and
translate to a prior on \code{alpha} that is approximately uniform.
\item The default prior for \code{mu_inf} spans the entire range of the score.
\item The default prior for \code{sigma_inf} translates to a range in the distribution of \code{y_inf} to be at most \code{max_score}.
\item The model is naive as it is trained with a non-truncated.
\item For more details see the \url{vignette}.
}
}
\section{Model's parameters}{


Population parameters:
\itemize{
\item \code{sigma}: Standard deviation of the autoregression
\item \code{mu_logit_alpha}: Population mean of the logit of \code{alpha}
\item \code{sigma_logit_alpha}: Population standard deviation of the logit of \code{alpha}
\item \code{mu_inf}: Population mean of \code{y_inf}
\item \code{sigma_inf}: Population standard deviation of \code{y_inf}
}

Patient-dependent parameters:
\itemize{
\item \code{alpha}: Autocorrelation parameter
\item \code{y_inf}: Autoregression mean
\item \code{b}: Intercept
}

Other parameters:
\itemize{
\item \code{y_mis}: Missing values
}

See \code{list_parameters(model = "MixedAR1")}.
}

