% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_doc.R
\name{OrderedMRW}
\alias{OrderedMRW}
\title{Ordered Logistic Multivariate Random Walk model}
\arguments{
\item{max_score}{Maximum value that the scores can take}

\item{D}{Number of components}

\item{independent_components}{Whether to have diagonal correlations matrices or not}

\item{prior}{Named list of the model's priors.
It uses the default priors (see \code{\link[=default_prior]{default_prior()}}) if \code{NULL} and for the parameters that are not provided.}
}
\description{
This is a state-space model defined by an ordered logistic measurement error distribution
and a latent multivariate random walk.
}
\details{
Details of the model are available in \strong{TBC}.
}
\section{Parameters}{


The parameters are essentially the same as the \link{OrderedRW} model,
except that they one extra dimension corresponding to the component.
In addition, there are two news population parameters:
\itemize{
\item \code{Omega}: Correlation matrix of the latent dynamic
\item \code{Sigma}: Covariance matrix of the latent dynamic
\item \code{Omega0}: Correlation matrix of the initial condition (cf. \code{logit_y0})
}

And one extra observation (patient- and time-dependent) parameter:
\itemize{
\item \code{ytot_rep}, which corresponds to the posterior replications of the sum of all components
}

See \code{list_parameters(model = "OrderedMRW")} for more details.
}

\section{Priors}{

The priors are passed as a named list with elements
\code{delta}, \code{sigma_meas}, \code{sigma_lat}, \code{Omega}, \code{Omega0}, \code{mu_y0} and \code{sigma_y0}
specifying priors for the corresponding parameters.

\code{Omega} and \code{Omega0} follow a LKJ distribution with shape parameter \code{eta} > 0.
The shape parameter \code{eta} can be interpreted like the shape parameter of a symmetric beta distribution.
When \code{eta} is 1, this corresponds to a "uniform" distribution over correlation matrices.
\verb{eta >> 1} converges to identity matrix.
\itemize{
\item \code{Omega ~ LJJ(eta)}
\item \code{Omega0 ~ LKJ(eta)}
}

Then, the priors for the remaining parameters are defined for each component,
and are the same as the priors for the univariate model \code{OrderedRW}.

The element \code{delta} should be a matrix with \code{D} rows and \code{max_score - 1} columns,
the d-th row corresponding to a vector of positive values X1,
defining the priors for the d-th component, \code{delta[d]}, such as:
\itemize{
\item \code{delta[d] ~ dirichlet(X1)}
}

The elements \code{sigma_meas}, \code{sigma_lat}, \code{mu_y0} and \code{sigma_y0}
should be matrices with 2 rows and \code{D} columns,
where the d-th column {x1 x2} correspond to the prior for the d-th component, with x2 > 0 and:
\itemize{
\item \code{sigma_meas[d] ~ lognormal(x1, x2)}
\item \code{sigma_meas[d] / max_score ~ lognormal(x1, x2)}
\item \code{sigma_lat[d] / max_score ~ lognormal(x1, x2)}
\item \code{mu_y0[d] ~ normal(x1, x2)}
\item \verb{sigma_y0[d] ~ normal+(x1, x2)}
}

NB: For \code{sigma_y0}, usually x1=0 to define a half-normal distribution
since the parameters are constrained to be positive.
}

\section{Default priors}{

\itemize{
\item The default priors for each component are the same as the default priors for \link{OrderedRW}.
\item The default priors for \code{Omega} and \code{Omega0} is a LKJ distribution with a shape parameter of 1.
}
}

\examples{
EczemaMVModel("OrderedMRW", max_score = 10, D = 5)
}
