% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plot_latent_OrderedRW}
\alias{plot_latent_OrderedRW}
\title{Plot the evolution of the expected latent score of the OrderedRW model}
\usage{
plot_latent_OrderedRW(fit, id, patient_id)
}
\arguments{
\item{fit}{Stanfit object}

\item{id}{Dataframe linking index in fit to (Patient, Time) pairs, cf. output from \code{\link[=get_index]{get_index()}}}

\item{patient_id}{Patient ID}
}
\value{
Ggplot
\itemize{
\item Horizontal lines correspond to the expected cut-offs
\item Ribbons correspond to the CI of a logistic distribution
}
}
\description{
Plot the evolution of the expected latent score of the OrderedRW model
}
