% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal.R
\name{prepare_data_lgtd}
\alias{prepare_data_lgtd}
\title{Prepares the data list to pass to the Stan sampler for longitudinal models}
\usage{
prepare_data_lgtd(train, test = NULL, max_score, discrete)
}
\arguments{
\item{train}{Training dataframe}

\item{test}{Testing dataframe}

\item{max_score}{Maximum value that the score can take}

\item{discrete}{Whether to use a discrete normal distribution (only relevant for testing)}
}
\value{
List to serve as input to the Stan sampler.
The list later needs to be appended by the item "discrete" for the RW model, "run" for the other models and
priors (see \code{add_prior()}).
}
\description{
Used internally.
}
\section{Dataframe format}{

\code{train} and \code{test} should have the columns \code{Patient} (patient ID), \code{Time} (timepoint) and \code{Score} (score to model).
\code{Patient} should take integer values between 1 and the number of patients in the training set.
\code{Time} should take integer (so discrete) values and starts with one for every patient.
\code{Score} should take values between 0 and max_score.
Missing values are not allowed (but Time values are not necessarily consecutive,
for example if Score at t=5 is missing, but not at t=4 and t=6, just remove t=5).
}

