% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC.R
\name{prepare_data_MC}
\alias{prepare_data_MC}
\title{Prepare data to pass to the Stan sampler for MC model}
\usage{
prepare_data_MC(train, test = NULL, K)
}
\arguments{
\item{train}{Training dataframe}

\item{test}{Testing dataframe}

\item{K}{Number of categories}
}
\description{
Prepare data to pass to the Stan sampler for MC model
}
\details{
\itemize{
\item \code{train} and \code{test} should have columns \code{y0} (for the current state), \code{y1} (for the next state) and
\code{dt} (for the time delay between states).
\item \code{y0} and \code{y1} should take integer values between 1 and K.
\item \code{dt} should take integer values greater than or equal to 1.
}

Missing values are not allowed.
}
