% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MV_model_class.R
\name{prepare_standata.EczemaMVModel}
\alias{prepare_standata.EczemaMVModel}
\title{Prepare data list to pass to the Stan sampler}
\usage{
\method{prepare_standata}{EczemaMVModel}(model, train, test = NULL, ...)
}
\arguments{
\item{model}{Object}

\item{train}{Training dataframe}

\item{test}{Testing dataframe}

\item{...}{Arguments to pass to other methods}
}
\value{
Named list
}
\description{
Prepare data list to pass to the Stan sampler
}
