% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R, R/prepare_standata.R
\name{prepare_standata}
\alias{prepare_standata}
\alias{prepare_data_lgtd}
\alias{prepare_standata.EczemaModel}
\alias{prepare_standata.MC}
\title{Prepare the data list to pass to the Stan sampler}
\usage{
prepare_standata(model, train, test, ...)

prepare_data_lgtd(train, test = NULL, max_score, discrete)

\method{prepare_standata}{EczemaModel}(model, train, test = NULL, ...)

\method{prepare_standata}{MC}(model, train, test = NULL, ...)
}
\arguments{
\item{model}{Object}

\item{train}{Training dataframe (details of the format in \code{\link[=EczemaFit]{EczemaFit()}})}

\item{test}{Testing dataframe (details of the format in \code{\link[=EczemaFit]{EczemaFit()}})}

\item{...}{Arguments to pass to other methods}

\item{max_score}{Maximum value that the score can take}

\item{discrete}{Whether to use a discrete normal distribution (only relevant for testing)}
}
\value{
List to serve as input to the Stan sampler.
The list is usually incomplete needs to be optional parameters, such as:
\itemize{
\item \code{run} (binary, for main and MC models, indicating whether to evaluate the likelihood)
\item \code{discrete} (binary, for RW model, indicating whether the predictions should be discretised)
}
}
\description{
Used internally.
}
\details{
\itemize{
\item \code{prepare_data_lgtd} is helps build \code{prepare_standata.EczemaModel} and is kept for compatibility reasons.
The list that it outputs does not include the priors.
}
}
