% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R
\name{print.EczemaModel}
\alias{print.EczemaModel}
\title{Print model}
\usage{
\method{print}{EczemaModel}(x, digits = 2, ...)
}
\arguments{
\item{x}{Object of class EczemaModel}

\item{digits}{Number of significant digits to print}
}
\value{
None
}
\description{
Print model
}
