% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MV_model_class.R
\name{print.EczemaMVModel}
\alias{print.EczemaMVModel}
\title{Print model}
\usage{
\method{print}{EczemaMVModel}(x, digits = 2, ...)
}
\arguments{
\item{x}{Object}

\item{digits}{Number of significants digits to print}

\item{...}{Arguments to pass to other methods}
}
\value{
None
}
\description{
Print model
}
