% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prior.R
\name{print_distribution}
\alias{print_distribution}
\title{Print distribution}
\usage{
print_distribution(parameter_name, distribution_name, arguments, digits = 2)
}
\arguments{
\item{parameter_name}{Name of the parameter}

\item{distribution_name}{Name of the distribution}

\item{arguments}{Arguments of the distribution}

\item{digits}{Number of significant digits to print}
}
\value{
None
}
\description{
Used internally in \code{\link[=print_prior]{print_prior()}}.
}
\examples{
print_distribution("x", "normal", c(0, 1))
}
