% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prior.R
\name{print_prior.BinMRW}
\alias{print_prior.BinMRW}
\title{Print prior of an OrderedMRW model}
\usage{
\method{print_prior}{BinMRW}(model, digits = 2, ...)
}
\arguments{
\item{model}{Object of class "OrderedMRW"}

\item{digits}{Number of significant digits to print}

\item{...}{Arguments to pass to other methods}
}
\value{
None
}
\description{
Print prior of an OrderedMRW model
}
