% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R, R/model_prior.R
\name{print_prior}
\alias{print_prior}
\alias{print_distribution}
\title{Print prior distribution}
\usage{
print_prior(model, ...)

print_distribution(parameter_name, distribution_name, arguments, digits = 2)
}
\arguments{
\item{model}{Object}

\item{...}{Arguments to pass to other methods}

\item{parameter_name}{Name of the parameter}

\item{distribution_name}{Name of the distribution}

\item{arguments}{Arguments of the distribution (numeric vector)}

\item{digits}{Number of significant digits to print (cf. \code{\link[base:signif]{base::signif()}})}
}
\value{
None
}
\description{
Used internally in the \code{print.EczemaModel} method.
}
\details{
\code{print_prior} usually calls \code{print_distribution} with the additional argument \code{digits}
(except for the base method for \code{EczemaModel} object).
}
\examples{
model <- EczemaModel("BinRW", max_score = 10)
print_prior(model)
print_prior(model, digits = 5)
print_distribution("x", "normal", c(0, 1))
}
