% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R
\name{replace_prior}
\alias{replace_prior}
\title{Replace prior}
\usage{
replace_prior(x, prior = NULL)
}
\arguments{
\item{x}{EczemaModel object}

\item{prior}{Named list of the model's prior to replace. If NULL, the prior stays the same}
}
\value{
Object of the same class as \code{x}
}
\description{
Used internally to overwrite default prior in the constructor.
Beware that the validity of the new prior is not tested in this function,
you may want to call \code{\link[=validate_prior]{validate_prior()}} after using this function.
}
\examples{
model <- EczemaModel("OrderedRW", max_score = 5)
print(model)
replace_prior(model, prior = list(sigma = c(0, 1)))
}
