% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{sample_prior_continuous}
\alias{sample_prior_continuous}
\title{Sample prior for continuous models}
\usage{
sample_prior_continuous(
  N_patient = 1,
  t_max = c(2),
  max_score,
  model = c("RW", "Smoothing", "AR1", "MixedAR1"),
  prior = default_prior(model, max_score),
  ...
)
}
\arguments{
\item{N_patient}{Number of patients}

\item{t_max}{Vector of size N_patient indicating the time-series length}

\item{max_score}{Maximum value that the score can take}

\item{model}{Model name}

\item{prior}{A named list of corresponding to parameters' prior.}

\item{...}{Arguments to be pass to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}
}
\value{
Stanfit object
}
\description{
Sample prior for continuous models
}
\seealso{
\code{\link[=sample_prior_RW]{sample_prior_RW()}}, \code{\link[=sample_prior_Smoothing]{sample_prior_Smoothing()}}, \code{\link[=sample_prior_AR1]{sample_prior_AR1()}}, \code{\link[=sample_prior_MixedAR1]{sample_prior_MixedAR1()}}
}
