% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{sample_prior_discrete}
\alias{sample_prior_discrete}
\title{Sample prior for discrete models}
\usage{
sample_prior_discrete(
  N_patient = 1,
  t_max = c(2),
  max_score,
  model = c("BinRW", "BinMC", "OrderedRW", "RW"),
  prior = default_prior(model, max_score),
  ...
)
}
\arguments{
\item{N_patient}{Number of patients}

\item{t_max}{Vector of size N_patient indicating the time-series length}

\item{max_score}{Maximum value that the score can take}

\item{model}{Model name}

\item{prior}{A named list of corresponding to parameters' prior.}

\item{...}{Arguments to be pass to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}
}
\value{
Stanfit object
}
\description{
Except Markov Chain model
}
\seealso{
\code{\link[=sample_prior_BinRW]{sample_prior_BinRW()}}, \code{\link[=sample_prior_BinMC]{sample_prior_BinMC()}}, \code{\link[=sample_prior_OrderedRW]{sample_prior_OrderedRW()}}, \code{\link[=sample_prior_RW]{sample_prior_RW()}}
}
