% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoothing.R
\name{Smoothing}
\alias{Smoothing}
\alias{default_prior_Smoothing}
\alias{fit_Smoothing}
\alias{sample_prior_Smoothing}
\title{Exponential smoothing model}
\usage{
default_prior_Smoothing()

fit_Smoothing(
  train,
  test = NULL,
  max_score,
  prior = default_prior_Smoothing(),
  ...
)

sample_prior_Smoothing(
  N_patient = 1,
  t_max = c(2),
  max_score,
  prior = default_prior_Smoothing(),
  ...
)
}
\arguments{
\item{train}{Training dataframe (details of the format in \code{\link[=prepare_data_lgtd]{prepare_data_lgtd()}})}

\item{test}{Testing dataframe (details of the format in \code{\link[=prepare_data_lgtd]{prepare_data_lgtd()}})}

\item{max_score}{Maximum value that the score can take}

\item{prior}{A named list with elements \code{sigma} and \code{tau} specifying priors for the corresponding parameters:
\itemize{
\item \code{sigma / max_score ~ normal(x1, x2)}.
The element \code{sigma} of the list should be a vector of length 2 containing x1 and x2.
NB: usually x1=0 to define a half-normal distribution (sigma is constraint to be positive) and
x2 should be positive.
\item \code{tau ~ lognormal(x1, x2)}.
The element \code{tau} of the list should be a vector of length 2 containing x1 and x2.
x2 should be positive.
}}

\item{...}{Arguments to be pass to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}

\item{N_patient}{Number of patients}

\item{t_max}{Vector of size N_patient indicating the time-series length}
}
\value{
\itemize{
\item \code{default_prior_Smoothing}: named list defining the default priors (see details)
\item \code{fit_Smoothing}: Stanfit object
\item \code{sample_prior_Smoothing}: Stanfit object
}
}
\description{
\itemize{
\item \code{default_prior_Smoothing} returns the default prior
\item \code{fit_Smoothing} fits the exponential smoothing model
\item \code{sample_prior_Smoothing} samples the prior predictive distribution
}
}
\details{
\itemize{
\item The default prior for \code{sigma} translates to a width of the predictive distribution to be at most \code{max_score}.
\item The default prior for \code{tau} assumes it could range from less a 1 to 100 (time units).
\item The model is naive as it is trained with a non-truncated
\item For more details see the \href{https://ghurault.github.io/EczemaPred/articles/ContinuousModels.html}{vignette}.
}
}
\section{Model's parameters}{

\itemize{
\item \code{sigma}: Standard deviation of the random walk
\item \code{alpha}: Smoothing factor
\item \code{tau}: Time constant associated with the smoothing factor
\item \code{y_mis}: Missing values
}

See \code{list_parameters(model = "Smoothing")}.
}

