% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prior.R
\name{validate_prior.BinMRW}
\alias{validate_prior.BinMRW}
\title{Check that the prior of an "BinMRW" model is correct}
\usage{
\method{validate_prior}{BinMRW}(model, ...)
}
\arguments{
\item{model}{Object of class "BinMRW"}

\item{...}{Arguments to pass to other methods}
}
\value{
NULL if the prior is correct, an error message otherwise.
}
\description{
Check that the prior of an "BinMRW" model is correct
}
