% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R
\name{validate_prior}
\alias{validate_prior}
\title{Check the prior of an EczemaModel is correct}
\usage{
validate_prior(model, ...)
}
\arguments{
\item{model}{Object}

\item{...}{Arguments to pass to other methods}
}
\value{
NULL if all statements are TRUE, otherwise an error message
}
\description{
Used internally when constructing an EczemaModel.
}
\seealso{
\code{\link[base:stopifnot]{base::stopifnot()}}
}
