% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{add_predictions}
\alias{add_predictions}
\alias{add_metrics}
\title{Append lpd, (C)RPS and predictive samples to testing dataframe}
\usage{
add_predictions(
  df,
  fit,
  discrete = TRUE,
  include_samples = FALSE,
  n_samples = NULL
)

add_metrics(df, fit, discrete = TRUE)
}
\arguments{
\item{df}{Dataframe.
For \code{add_metrics}, when \code{discrete = FALSE}, it must contain a column "Score".}

\item{fit}{Stanfit object}

\item{discrete}{Whether to estimate a discrete or continuous forecast.
For a discrete forecast, the RPS will be computed and the CRPS for a continuous forecast.}

\item{include_samples}{Whether to return samples from the historical forecast in the output}

\item{n_samples}{If include_samples=TRUE, how many samples to return. Default (=NULL) to all samples.}
}
\value{
Dataframe df appended by the columns "lpd", "RPS" (or CRPS if discrete=FALSE) and optionally "Samples"
}
\description{
Append lpd, (C)RPS and predictive samples to testing dataframe
}
