% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_doc.R
\name{AR1}
\alias{AR1}
\title{Autoregressive model (order 1)}
\description{
Autoregressive model (order 1)
}
\details{
\itemize{
\item Details of the model are available in the \href{#}{paper}.
\item The model is naive as it is trained with a non-truncated distribution
\item For more details see the \href{https://ghurault.github.io/EczemaPred/articles/ContinuousModels.html}{vignette}.
}

\itemize{
\item The model is naive as it is trained with a non-truncated distribution
\item For more details see the \href{https://ghurault.github.io/EczemaPred/articles/ContinuousModels.html}{vignette}.
}
}
\section{Parameters}{

\itemize{
\item \code{sigma}: Standard deviation of the autoregression
\item \code{alpha}: Autocorrelation parameter
\item \code{b}: Intercept
\item \code{y_inf}: Autoregression mean
\item \code{y_mis}: Missing values
}

See \code{list_parameters(model = "AR1")}.
}

\section{Priors}{

The priors are passed as a named list with elements \code{sigma}, \code{y_inf} and \code{alpha}
specifying priors for the corresponding parameters.
Each element of the list should be a vector of length 2, containing values for x1 and x2, x2 > 0, such as:
\itemize{
\item \verb{sigma / max_score ~ normal+(x1, x2)}.
\item \code{y_inf / max_score ~ normal(x1, x2)}.
\item \code{alpha ~ beta(x1, x2)}.
}

NB: For \code{sigma}, usually x1=0 to define a half-normal distribution
since the parameter is constrained to be positive.
NB: For \code{alpha}, both \code{x1} and \code{x2} must be positive.
}

\section{Default priors}{

\itemize{
\item The default prior for \code{sigma} translates to a width of the predictive distribution to be at most \code{max_score}.
\item The default prior for \code{y_inf} covers the full range of the score.
\item The default prior for \code{alpha} is uniform in 0-1.
}
}

\examples{
EczemaModel("AR1", max_score = 100)
}
