% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_doc.R
\name{BinMC}
\alias{BinMC}
\title{Binomial Markov Chain model}
\arguments{
\item{max_score}{Maximum value that the score can take}

\item{prior}{Named list of the model's priors. If \code{NULL}, uses the default prior for the model (see \code{\link[=default_prior]{default_prior()}}).}
}
\description{
This is a state-space model defined by a Binomial measurement error and a latent Markov Chain.
For more details see the BinRW \href{https://ghurault.github.io/EczemaPred/articles/BinRW.html}{vignette}.
}
\details{
Details of the model are available in the \href{#}{paper}.
}
\section{Parameters}{

\subsection{Population parameters:}{
\itemize{
\item \code{sigma}: Standard deviation of the evolution of \code{ss1}
\item \code{mu_logit_p10}: Population logit mean of \code{p10}
\item \code{sigma_logit_p10}: Population logit standard deviation of \code{p10}
}
}

\subsection{Patient-dependent parameters:}{
\itemize{
\item \code{p10}: Probability of transitioning from state 1 to state 0
\item \code{logit_p10}: logit of \code{p10}
\item \code{logit_tss1_0}: Initial condition of the \code{logit(ss1 * (1 + p10))}
}
}

\subsection{Observation-dependent (patient- and time-dependent) parameters:}{
\itemize{
\item \code{p01}: Probability of transitioning from state 0 to state 1
\item \code{lambda}: Mobility of the Markov Chain (eigen value of the transition matrix)
\item \code{ss1}: Steady state probability of state 1
\item \code{y_lat}: Latent score (probability)
}

See \code{list_parameters(model = "BinMC")} for more details.
}
}

\section{Priors}{

The priors are passed as a named list with elements \code{sigma}, \code{mu_logit_p10} and \code{sigma_logit_p10}
specifying priors for the corresponding parameters.
Each element of the list should be a vector of length 2, containing values for x1 and x2, x2 > 0, such as:
\itemize{
\item \verb{sigma ~ normal+(x1, x2)}
\item \code{mu_logit_p10 ~ normal(x1, x2)}
\item \verb{sigma_logit_p10 ~ normal+(x1, x2)}
\item \code{logit_tss1_0 ~ normal(x1, x2)}
}

NB: For \code{sigma} and \code{sigma_logit_p10}, usually x1=0 to define a half-normal distribution
since the parameter is constrained to be positive.
}

\section{Default priors}{

\itemize{
\item The default prior for \code{sigma} translates to an odd ratio increment of at most 5 (~ 2 * upper bound of prior).
\item The default priors for \code{mu_logit_p10} and \code{sigma_logit_p10} translate to an approximately uniform prior on \code{p10}.
\item The prior for the initial condition of \code{ss1} is hard coded and a function of \code{p10}.
}
}

\examples{
EczemaModel("BinMC", max_score = 100)
}
