% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_doc.R
\name{BinRW}
\alias{BinRW}
\title{Binomial random walk model}
\arguments{
\item{max_score}{Maximum value that the score can take}

\item{prior}{Named list of the model's priors. If \code{NULL}, uses the default prior for the model (see \code{\link[=default_prior]{default_prior()}}).}
}
\description{
This is a state-space model defined by a Binomial measurement error and a latent random walk.
For more details see the \href{https://ghurault.github.io/EczemaPred/articles/BinRW.html}{vignette}.
}
\details{
Details of the model are available in the \href{#}{paper}.
}
\section{Parameters}{


Population parameters:
\itemize{
\item \code{sigma}: Standard deviation of the random walk (in logit scale)
\item \code{mu_logit_y0}: Population mean of the initial condition (in logit scale)
\item \code{sigma_logit_y0}: Population standard deviation of the initial condition (logit scale)
}

Patient-dependent parameters:
\itemize{
\item \code{logit_y0}: Logit of the initial condition
}

Observation-dependent (patient- and time-dependent) parameters:
\itemize{
\item \code{y_lat}: Latent score (probability)
\item \code{logit_lat}: logit of \code{y_lat}
}

See \code{list_parameters(model = "BinRW")}.
}

\section{Priors}{

The priors are passed as a named list with elements \code{sigma}, \code{mu_logit_y0} and \code{sigma_logit_y0}
specifying priors for the corresponding parameters.
Each element of the list should be a vector of length 2, containing values for x1 and x2, x2 > 0, such as:
\itemize{
\item \verb{sigma ~ normal+(x1, x2)}.
\item \code{mu_logit_y0 ~ normal(x1, x2)}
\item \verb{sigma_logit_y0 ~ normal+(x1, x2)}
}

NB: For \code{sigma} and \code{sigma_logit_y0}, usually x1=0 to define a half-normal distribution
since the parameters are constrained to be positive.
}

\section{Default priors}{

\itemize{
\item The default priors do not depend on \code{max_score}, and are printed in the examples section.
\item The default prior for \code{sigma} translates to an odd ratio increment of at most 5 (~ 2 * upper bound of prior).
\item The default priors for \code{mu_logit_y0} and \code{sigma_logit_y0} translates to an approximately uniform prior on \code{y0}.
}
}

\examples{
EczemaModel("BinRW", max_score = 100)
}
