% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R, R/model_prior.R
\name{default_prior}
\alias{default_prior}
\alias{default_prior.character}
\alias{list_parameters.character}
\alias{list_parameters.BinRW}
\alias{default_prior.BinRW}
\title{Default prior}
\usage{
default_prior(model, ...)

\method{default_prior}{character}(model, max_score = 1, ...)

\method{list_parameters}{character}(model, ...)

\method{list_parameters}{BinRW}(model, main = TRUE)

\method{default_prior}{BinRW}(model)
}
\arguments{
\item{model}{Object}

\item{...}{Arguments to pass to other methods}
}
\value{
Named list of parameters' priors.
For more details, see the generic of the model class.
}
\description{
Default prior
}
\examples{
default_prior("BinRW")
list_parameters("BinRW")
list_parameters(EczemaModel("BinRW", max_score = 100))
default_prior(EczemaModel("BinRW", max_score = 10))
}
