% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R, R/model_prior.R
\name{default_prior}
\alias{default_prior}
\alias{default_prior.BinRW}
\title{Default prior}
\usage{
default_prior(model)

\method{default_prior}{BinRW}(model)
}
\arguments{
\item{model}{Object of class EczemaModel}
}
\value{
Named list of parameters' priors.
For more details, see the generic of the model class.
}
\description{
Default prior
}
\examples{
default_prior(EczemaModel("BinRW", max_score = 10))
}
