% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R, R/model_prior.R
\name{default_prior}
\alias{default_prior}
\alias{default_prior.character}
\alias{default_prior.BinRW}
\alias{default_prior.OrderedRW}
\alias{default_prior.BinMC}
\title{Default prior}
\usage{
default_prior(model, ...)

\method{default_prior}{character}(model, max_score = 1, ...)

\method{default_prior}{BinRW}(model)

\method{default_prior}{OrderedRW}(model)

\method{default_prior}{BinMC}(model)
}
\arguments{
\item{model}{Object}

\item{...}{Arguments to pass to other methods}
}
\value{
Named list of parameters' priors.
For more details, see the generic of the model class.
}
\description{
Default prior
}
\examples{
default_prior("BinRW")
default_prior(EczemaModel("BinRW", max_score = 10))
default_prior(EczemaModel("OrderedRW", max_score = 10))
default_prior(EczemaModel("BinMC", max_score = 100))
}
