% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R, R/model_prior.R
\name{default_prior}
\alias{default_prior}
\alias{default_prior.character}
\title{Default prior}
\usage{
default_prior(model, ...)

\method{default_prior}{character}(model, max_score = 1, K = 2, ...)
}
\arguments{
\item{model}{Object}

\item{...}{Arguments to pass to other methods}

\item{max_score}{Maximum value that the score can take}

\item{K}{Number of categories}
}
\value{
Named list of parameters' priors.
For more details, see the generic of the model class.
}
\description{
Default prior
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: The function creates an EczemaModel object and call the corresponding method.
}}

\examples{
default_prior(EczemaModel("BinRW", max_score = 10))
default_prior(EczemaModel("MC", K = 10))
default_prior("BinRW")

}
