% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R, R/model_inference.R
\name{EczemaFit}
\alias{EczemaFit}
\alias{EczemaFit.EczemaModel}
\title{Fit an EczemaModel}
\usage{
EczemaFit(model, train, test, ...)

\method{EczemaFit}{EczemaModel}(model, train, test = NULL, ...)
}
\arguments{
\item{model}{Object}

\item{train}{Training dataframe}

\item{test}{Testing dataframe}

\item{...}{Arguments to pass to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}
}
\value{
Stanfit object
}
\description{
Fit an EczemaModel
}
\section{Methods (by class)}{
\itemize{
\item \code{EczemaModel}: Method for EczemaModel object
}}

\examples{
\dontrun{
model <- EczemaModel("BinRW", max_score = 100)
train <- data.frame(Patient = 1, Time = 1:10, Score = rbinom(10, 100, .5))
EczemaFit(model, train)
}
}
