% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R, R/model_inference.R
\name{EczemaFit}
\alias{EczemaFit}
\alias{EczemaFit.EczemaModel}
\title{Fit an EczemaModel}
\usage{
EczemaFit(model, train, test, ...)

\method{EczemaFit}{EczemaModel}(model, train, test = NULL, ...)
}
\arguments{
\item{model}{Object of class EczemaModel}

\item{train}{Training dataframe}

\item{test}{Testing dataframe}

\item{...}{Arguments to pass to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}
}
\value{
Object of class stanfit
}
\description{
Fit an EczemaModel
}
\section{Methods (by class)}{
\itemize{
\item \code{EczemaModel}: Method for EczemaModel object
}}

