% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R
\name{EczemaModel}
\alias{EczemaModel}
\title{EczemaModel constructor}
\usage{
EczemaModel(
  model_name = c("BinRW", "OrderedRW", "BinMC", "RW", "Smoothing", "AR1", "MixedAR1"),
  max_score = NULL,
  K = NULL,
  discrete = TRUE,
  prior = NULL
)
}
\arguments{
\item{model_name}{Name of the model to create}

\item{max_score}{Maximum value that the score can take. Required for all models except "MC".}

\item{K}{Number of categories. Only required for "MC" model.}

\item{discrete}{Whether the model is discrete or not. Only required for "RW".}

\item{prior}{Named list of the model's priors. If \code{NULL}, uses the default prior for the model (see \code{\link[=default_prior]{default_prior()}}).}
}
\value{
An object (list) of class \code{model_name} and EczemaModel
}
\description{
EczemaModel constructor
}
\examples{
EczemaModel("BinRW", max_score = 10)
}
