% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EczemaPred-package.R
\docType{package}
\name{EczemaPred-package}
\alias{EczemaPred-package}
\alias{EczemaPred}
\title{The 'EczemaPred' package.}
\description{
A collection of models to serve as building blocks for predicting eczema severity.
}
\references{
\href{https://mc-stan.org}{Stan Development Team (2020). RStan: the R interface to Stan.} R package version 2.21.2.
}
