% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{extract_parameters}
\alias{extract_parameters}
\title{Extract parameters posterior summary statistics}
\usage{
extract_parameters(fit, pars = NULL, id = NULL, ...)
}
\arguments{
\item{fit}{Stanfit object}

\item{pars}{Named list of parameters to extract. See \code{\link[=list_parameters]{list_parameters()}}.
If \code{NULL}, it extracts all parameters of the model.}

\item{id}{Dataframe associating Index to (Patient, Time) pairs.
See \code{\link[=get_index]{get_index()}}.
\code{id} is not used when \code{pars = NULL} or \code{pars} does not contain an element \code{PatientTime}.}

\item{...}{Arguments to pass to \code{\link[HuraultMisc:summary_statistics]{HuraultMisc::summary_statistics()}}.}
}
\value{
Tibble dataframe containing posterior summary statistics.
See details in \code{\link[HuraultMisc:summary_statistics]{HuraultMisc::summary_statistics()}}. Additional columns Patient and Time if \code{id} is not NULL.
}
\description{
Extract parameters posterior summary statistics
}
\examples{
\dontrun{
model <- EczemaModel("BinRW", max_score = 100)
id <- get_index2(rpois(10, 20))
train <- id \%>\% mutate(Score = rbinom(nrow(.), 100, .5))
fit <- EczemaFit(model, train)
extract_parameters(fit, pars = list_parameters("BinRW"), id = id)
}
}
