% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_compiled_model}
\alias{get_compiled_model}
\title{Get compiled model}
\usage{
get_compiled_model(model)
}
\arguments{
\item{model}{Stan model name}
}
\value{
Compiled model (object to pass to \code{rstan::sampling})
}
\description{
Get compiled model
}
