% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_longitudinal.R
\name{get_index}
\alias{get_index}
\alias{get_index2}
\title{Associate (Patient, Time) pairs to corresponding index in the model}
\usage{
get_index(train, test = NULL)

get_index2(t_max)
}
\arguments{
\item{train}{Training dataframe}

\item{test}{Testing dataframe}

\item{t_max}{Vector indicating the length of each patient time-series}
}
\value{
Dataframe with columns Patient, Time, Index
}
\description{
Associate (Patient, Time) pairs to corresponding index in the model
}
\details{
These functions are not designed for use with the Markov Chain model (MC).
}
\examples{
library(dplyr)
id <- get_index2(t_max = rpois(10, 20))
df <- id \%>\% select(-Index) \%>\% slice_sample(prop = 0.9) \%>\% arrange(Patient, Time)
get_index(train = df)
}
