% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R, R/parameters.R
\name{list_parameters}
\alias{list_parameters}
\alias{list_parameters.BinRW}
\alias{list_parameters.BinMC}
\alias{list_parameters.OrderedMRW}
\alias{list_parameters.BinMRW}
\title{List available parameters}
\usage{
list_parameters(model, ...)

\method{list_parameters}{BinRW}(model, main = TRUE, ...)

\method{list_parameters}{BinMC}(model, main = TRUE, ...)

\method{list_parameters}{OrderedMRW}(model, full_names = FALSE, ...)

\method{list_parameters}{BinMRW}(model, full_names = FALSE, ...)
}
\arguments{
\item{model}{Object, usually the model name or an EczemaModel object}

\item{...}{Arguments to pass to other methods}

\item{main}{Whether to output the main parameters only (when applicable).}

\item{full_names}{(experimental, for multivariate models)
Whether to return the full names of multi-dimensional parameters (\code{Omega}, \code{Omega0} and \code{ct}).}
}
\value{
Named list of parameters names, grouped into broad categories:
\itemize{
\item \code{Population}: population parameters (i.e. patient- and time-independent)
\item \code{Patient}: patient-dependent parameters
\item \code{PatientTime}: patient- and time-dependent parameters (e.g. latent scores)
\item \code{Test}: parameters related to the test set
\item \code{Misc}: other parameters
}
}
\description{
List available parameters
}
\details{
See \link{MC}, \link{BinRW}, \link{BinMC}, \link{OrderedRW}, \link{RW}, \link{Smoothing}, \link{AR1} and \link{MixedAR1} for details about the model-specific parameters.
Other parameters are available across models:
\itemize{
\item \code{y_rep} correspond to posterior replications. To get the corresponding index, use \code{\link[=get_index]{get_index()}}.
\item \code{y_pred} is a subset of y_rep corresponding to test samples (size \code{N_test} equal to the number of observations in the test set).
\item \code{lpd} is the log predictive density of test samples (of size \code{N_test}).
\item \code{cum_err} is the cumulative error distribution, only available for discrete outcomes
(matrix with dimensions \code{N_test * (max_score  + 1)}).
}
}
\examples{
list_parameters("RW")
list_parameters(EczemaModel("RW", max_score = 100))
list_parameters(EczemaModel("BinRW", max_score = 100))
list_parameters(EczemaModel("BinMC", max_score = 100))
}
