% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_doc.R
\name{MC}
\alias{MC}
\title{Markov Chain model}
\arguments{
\item{K}{Number of states of the Markov Chain}

\item{prior}{Named list of the model's priors. If \code{NULL}, uses the default prior for the model (see \code{\link[=default_prior]{default_prior()}}).}
}
\description{
For more details see the \href{https://ghurault.github.io/EczemaPred/articles/MC.html}{Markov Chain vignette}.
}
\details{
Details of the model are available in the \href{#}{paper}.
}
\section{Parameters}{

\itemize{
\item \code{p}: matrix of size K * K where \code{p[i, j]} represents the transition probabilities from state i to state j.
}

See \code{list_parameters(model = "MC")}.
}

\section{Priors}{

The priors are passed as a named list with element \code{p}.
The transition probabilities from state i \code{p[i, ]} are assumed to follow a Dirichlet distribution.
The prior should be a matrix where each line correspond to the parameters of the Dirichlet distribution for \code{p[i, ]}.
}

\section{Default priors}{

The default prior for all \code{p[i, ]} is a symmetric uniform Dirichlet distribution (all concentration parameters are equal to 1).
}

\examples{
EczemaModel("MC", K = 5)
}
