% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC.R
\name{MC}
\alias{MC}
\alias{default_prior_MC}
\alias{prepare_data_MC}
\alias{fit_MC}
\alias{sample_prior_MC}
\title{Markov Chain model}
\usage{
default_prior_MC(K)

prepare_data_MC(train, test = NULL, K, prior = default_prior_MC(K))

fit_MC(train, test = NULL, K, prior = default_prior_MC(K), ...)

sample_prior_MC(
  train = data.frame(y0 = integer(), y1 = integer(), dt = integer()),
  K,
  prior = default_prior_MC(K),
  ...
)
}
\arguments{
\item{K}{Number of states of the Markov Chain}

\item{train}{Training dataframe}

\item{test}{Testing dataframe}

\item{prior}{A named list of corresponding to parameters' prior.
The transition probabilities from state i \code{p[i, ]} is assumed to follow a Dirichlet distribution.
The prior should be a matrix where each line correspond to the parameters of the Dirichlet distribution for \code{p[i, ]}.}

\item{...}{Arguments to be pass to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}
}
\value{
\itemize{
\item \code{default_prior_MC}: list corresponding to a uniform Dirichlet prior of concentration for each distribution \code{p[i, ]}
\item \code{prepare_data_MC}: list to serve as input to the Stan sampler.
\item \code{fit_MC}: Stanfit object
\item \code{sample_prior_MC}: Stanfit object
}
}
\description{
\itemize{
\item \code{default_prior_MC} returns the default prior
\item \code{prepare_data_MC} prepares the data list to pass to the Stan sampler
\item \code{fit_MC} fits the Markov Chain model
\item \code{sample_prior_MC} samples the prior predictive distribution
}
}
\details{
For more details see the \url{Markov Chain vignette}.
}
\section{Dataframe format}{

\itemize{
\item \code{train} and \code{test} should have columns \code{y0} (for the current state), \code{y1} (for the next state) and
\code{dt} (for the time delay between states).
\item \code{y0} and \code{y1} should take integer values between 1 and K.
\item \code{dt} should take integer values greater than or equal to 1.
}

Missing values are not allowed.
}

\section{Model's parameters}{

\itemize{
\item \code{p}: matrix of size K * K where \code{p[i, j]} represents the transition probabilities from state i to state j.
}

See \code{list_parameters(model = "MC")}.
}

