% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_doc.R
\name{OrderedRW}
\alias{OrderedRW}
\title{Ordered Logistic random walk model}
\arguments{
\item{max_score}{Maximum value that the score can take}

\item{prior}{Named list of the model's priors. If \code{NULL}, uses the default prior for the model (see \code{\link[=default_prior]{default_prior()}}).}
}
\description{
This is a state-space model defined by a Ordered logistic measurement error distribution and a latent random walk.
For more details see the BinRW \href{https://ghurault.github.io/EczemaPred/articles/BinRW.html}{vignette}.
}
\details{
Details of the model are available in the \href{#}{paper}.
}
\section{Parameters}{

\subsection{Population parameters:}{
\itemize{
\item \code{sigma}: Standard deviation of the random walk
\item \code{mu_y0}: Population mean of \code{y0} (initial condition).
\item \code{sigma_y0}: Population standard deviation of \code{y0} (initial condition).
\item \code{delta}: Difference between cutpoints (vector of length \code{max_score - 1})
\item \code{ct}: Cutpoints (vector of length \code{max_score})
\item \code{p0}: Probability distribution of the average patient at t0 (vector of length \code{max_score})
}
}

\subsection{Patient-dependent parameters:}{
\itemize{
\item \code{y0}: \code{y_lat} at t0.
}
}

\subsection{Observation-dependent (patient- and time-dependent) parameters:}{
\itemize{
\item \code{y_lat}: Latent score
}

See \code{list_parameters(model = "OrderedRW")} for more details.
}
}

\section{Priors}{

The priors are passed as a named list with elements \code{delta}, \code{sigma}, \code{mu_y0} and \code{sigma_y0}
specifying priors for the corresponding parameters.

The element \code{delta} should be a matrix with 2 rows and \code{max_score - 1} columns,
such as the i-th column is a vector with values x1 and x2, where x2 > 0 and
\verb{delta[i] ~ normal+(x1, x2)}.
The other parameters are normalised by the difference between the highest and lowest cutpoints (approx. the range of the score),
and their priors are defined by a vector of length 2, containing values for x1 and x2, x2 > 0, such as:
\itemize{
\item \verb{sigma ~ normal+(x1, x2)}
\item \code{mu_y0 ~ normal(x1, x2)}
\item \verb{sigma_y0 ~ normal+(x1, x2)}
}

NB: \code{delta}, \code{sigma} and \code{sigma_y0} are constrained to be positive so x1 are usually set to 0 to define a half-normal distribution.
}

\section{Default priors}{

\itemize{
\item The default prior for \code{delta} is set so that \code{delta} is less than the width of the logistic distribution.
\item The default prior for \code{sigma} assumes it would be to go to a state where \code{y = 0} is the most likely outcome to
a state where \code{y = M} in two transitions.
\item The default priors for \code{mu_y0} and \code{sigma_y0} have reasonable ranges and translate to an approximately uniform prior
over the range of the score for \code{y0}.
}
}

\examples{
EczemaModel("OrderedRW", max_score = 10)
}
