% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderedRW.R
\name{OrderedRW}
\alias{OrderedRW}
\alias{default_prior_OrderedRW}
\alias{fit_OrderedRW}
\alias{sample_prior_OrderedRW}
\title{Ordered Logistic random walk model}
\usage{
default_prior_OrderedRW(max_score)

fit_OrderedRW(
  train,
  test = NULL,
  max_score,
  prior = default_prior_OrderedRW(max_score),
  ...
)

sample_prior_OrderedRW(
  N_patient = 1,
  t_max = c(2),
  max_score,
  prior = default_prior_OrderedRW(max_score),
  ...
)
}
\arguments{
\item{max_score}{Maximum value that the score can take}

\item{train}{Training dataframe (details of the format in \code{\link[=prepare_data_lgtd]{prepare_data_lgtd()}})}

\item{test}{Testing dataframe (details of the format in \code{\link[=prepare_data_lgtd]{prepare_data_lgtd()}})}

\item{prior}{A named list with elements \code{delta}, \code{sigma}, \code{mu_y0} and \code{sigma_y0}
specifying priors for the corresponding parameters.

The element \code{delta} should be a matrix with 2 rows and \code{max_score - 1} columns,
such as the i-th column is a vector with values x1 and x2, where x2 > 0 and
\code{delta[i] ~ normal(x1, x2)}.
The other parameters are normalised by the difference between the highest and lowest cutpoints (approx. the range of the score),
and their priors are defined by a vector of length 2, containing values for x1 and x2, x2 > 0, such as:
\itemize{
\item \code{sigma ~ normal(x1, x2)}
\item \code{mu_y0 ~ normal(x1, x2)}
\item \code{sigma_y0 ~ normal(x1, x2)}
}

NB: \code{delta}, \code{sigma} and \code{sigma_y0} are constrained to be positive so x1 are usually set to 0 to define a half-normal distribution.}

\item{...}{Arguments to be pass to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}

\item{N_patient}{Number of patients}

\item{t_max}{Vector of size N_patient indicating the time-series length}
}
\value{
\itemize{
\item \code{default_prior_OrderedRW}: named list containing the default prior for \code{delta}
\item \code{fit_OrderedRW}: Stanfit object
\item \code{sample_prior_OrderedRW}: Stanfit object
}
}
\description{
This is a state-space model defined by a Ordered logistic measurement error distribution and a latent random walk.
}
\details{
\itemize{
\item \code{default_prior_OrderedRW} returns the default prior
\item \code{fit_OrderedRW} fits the Ordered logistic random walk model
\item \code{sample_prior_OrderedRW} samples the prior predictive distribution
\item The default prior for \code{delta} is set so that \code{delta} is less than the width of the logistic distribution.
\item The default prior for \code{sigma} assumes it would be to go to a state where \code{y = 0} is the most likely outcome to
a state where \code{y = M} in two transitions.
\item The default priors for \code{mu_y0} and \code{sigma_y0} have reasonable ranges and translate to an approximately uniform prior
over the range of the score for \code{y0}.
\item For more details see the \url{vignette}.
}
}
\section{Model's parameters}{


Population parameters:
\itemize{
\item \code{sigma}: Standard deviation of the random walk
\item \code{mu_y0}: Population mean of \code{y0} (initial condition).
\item \code{sigma_y0}: Population standard deviation of \code{y0} (initial condition).
\item \code{delta}: Difference between cutpoints (vector of length \code{max_score - 1})
\item \code{ct}: Cutpoints (vector of length \code{max_score})
\item \code{p0}: Probability distribution of the average patient at t0 (vector of length \code{max_score})
}

Patient-dependent parameters:
\itemize{
\item \code{y0}: \code{y_lat} at t0.
}

Observation-dependent (patient- and time-dependent) parameters:
\itemize{
\item \code{y_lat}: Latent score
}

See \code{list_parameters(model = "OrderedRW")}.
}

