% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ppc.R
\name{plot_ppc}
\alias{plot_ppc}
\alias{plot_post_traj_pmf}
\alias{plot_post_traj_fanchart}
\alias{plot_ppc_traj_pmf}
\alias{plot_ppc_traj_fanchart}
\title{Plot posterior predictive trajectory}
\usage{
plot_post_traj_pmf(obj, id, patient_id, max_score = NA, max_scale = NA)

plot_post_traj_fanchart(
  obj,
  id,
  patient_id,
  max_score = NA,
  interval = c("eti", "hdi"),
  CI_level = seq(0.1, 0.9, 0.1)
)

plot_ppc_traj_pmf(obj, train, test, patient_id, max_score = NA, max_scale = NA)

plot_ppc_traj_fanchart(
  obj,
  train,
  test,
  patient_id,
  max_score = NA,
  interval = c("eti", "hdi"),
  CI_level = seq(0.1, 0.9, 0.1)
)
}
\arguments{
\item{obj}{Stanfit object or matrix of replications, with rows corresponding to samples and
columns corresponding to variables (there should be \code{nrow(id)} columns)}

\item{id}{Dataframe linking index in obj to (Patient, Time) pairs, cf. output from \code{\link[=get_index]{get_index()}}}

\item{patient_id}{Patient ID}

\item{max_score}{Maximum value that the score can take.
Can be NA, in that case the support of the pmf is estimated from the data (cf. \code{\link[HuraultMisc:extract_distribution]{HuraultMisc::extract_distribution()}}).}

\item{max_scale}{Maximum value that the legend display.
If NA, this chosen automatically}

\item{interval}{Type of confidence of interval to display, one of "eti" for equal-tailed intervals and
"hdi" for highest density interval.}

\item{CI_level}{Vector of confidence level to plot for the fanchart}

\item{train}{Training dataset used to obtain the fit}

\item{test}{Testing dataset used to obtain the fit (can be NULL)}
}
\value{
Ggplot
}
\description{
Plot as a probability mass function (suffix \verb{_pmf}) or
as a fanchart (stacked confidence interval, suffix \verb{_fanchart}).
}
\details{
\itemize{
\item plot_post_traj_* plots the posterior predictive trajectory only
\item plot_ppc_traj_* overlays the observed trajectory to the posterior predictive trajectory
}
}
