% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC.R
\name{plot_transition_MC}
\alias{plot_transition_MC}
\title{Markov Chain expected transition matrix}
\usage{
plot_transition_MC(fit)
}
\arguments{
\item{fit}{Stanfit object corresponding to the Markov Chain model}
}
\value{
Ggplot
}
\description{
Markov Chain expected transition matrix
}
