% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R
\name{print.EczemaModel}
\alias{print.EczemaModel}
\title{Print model}
\usage{
\method{print}{EczemaModel}(model, digits = 2)
}
\arguments{
\item{model}{Object of class EczemaModel}

\item{digits}{Number of significant digits to print}
}
\value{
Invisible
}
\description{
Print model
}
