% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R
\name{print_prior}
\alias{print_prior}
\title{Print prior distribution}
\usage{
print_prior(model, ...)
}
\arguments{
\item{model}{Object}

\item{...}{Arguments to pass to other methods}

\item{digits}{Number of significant digits to print}
}
\value{
None
}
\description{
Used internally in the \code{print.EczemaModel} method.
}
\examples{
model <- EczemaModel("BinRW", max_score = 10)
print_prior(model)
print_prior(model, digits = 5)
}
