% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R, R/model_prior.R
\name{print_prior}
\alias{print_prior}
\alias{print_prior.BinRW}
\alias{print_prior.OrderedRW}
\alias{print_prior.BinMC}
\alias{print_prior.RW}
\alias{print_prior.Smoothing}
\alias{print_prior.AR1}
\alias{print_prior.MixedAR1}
\alias{print_prior.MC}
\title{Print prior distribution}
\usage{
print_prior(model, ...)

\method{print_prior}{BinRW}(model, digits = 2)

\method{print_prior}{OrderedRW}(model, digits = 2)

\method{print_prior}{BinMC}(model, digits = 2)

\method{print_prior}{RW}(model, digits = 2)

\method{print_prior}{Smoothing}(model, digits = 2)

\method{print_prior}{AR1}(model, digits = 2)

\method{print_prior}{MixedAR1}(model, digits = 2)

\method{print_prior}{MC}(model, digits = 2)
}
\arguments{
\item{model}{Object}

\item{...}{Arguments to pass to other methods}
}
\value{
None
}
\description{
Used internally in the \code{print.EczemaModel} method.
}
