% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R, R/model_prior.R
\name{print_prior}
\alias{print_prior}
\alias{print_prior.BinRW}
\title{Print prior distribution}
\usage{
print_prior(model, ...)

\method{print_prior}{BinRW}(model, digits = 2)
}
\arguments{
\item{model}{Object}

\item{...}{Arguments to pass to other methods}
}
\value{
Invisible
}
\description{
Print prior distribution
}
