% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R, R/model_inference.R
\name{sample_prior}
\alias{sample_prior}
\alias{sample_prior.EczemaModel}
\title{Prior predictive distribution}
\usage{
sample_prior(model, ...)

\method{sample_prior}{EczemaModel}(model, N_patient = 1, t_max = c(2), ...)
}
\arguments{
\item{model}{Object}

\item{...}{Arguments to pass to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}

\item{N_patient}{Number of patients to simulate}

\item{t_max}{Vector of size \code{N_patient} indicating the time-series length of each patient}
}
\value{
Object of class stanfit
}
\description{
Prior predictive distribution
}
\section{Methods (by class)}{
\itemize{
\item \code{EczemaModel}: Method for EczemaModel object
}}

\examples{
\dontrun{
model <- EczemaModel("BinRW", max_score = 100)
sample_prior(model)
}
}
