% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_doc.R
\name{Smoothing}
\alias{Smoothing}
\title{Exponential smoothing model}
\arguments{
\item{max_score}{Maximum value that the score can take}

\item{discrete}{Whether to use a discrete normal distribution.
This will be used to check whether the data is discrete or not, and for rounding predictions (cf. testing).}

\item{prior}{Named list of the model's priors. If \code{NULL}, uses the default prior for the model (see \code{\link[=default_prior]{default_prior()}}).}
}
\description{
Exponential smoothing model
}
\details{
\itemize{
\item Details of the model are available in the \href{#}{paper}.
\item The model is naive as it is trained with a non-truncated distribution.
\item For more details see the \href{https://ghurault.github.io/EczemaPred/articles/ContinuousModels.html}{vignette}.
}
}
\section{Parameters}{

\itemize{
\item \code{sigma}: Standard deviation of the random walk
\item \code{alpha}: Smoothing factor
\item \code{tau}: Time constant associated with the smoothing factor
\item \code{y_mis}: Missing values
}

See \code{list_parameters(model = "Smoothing")} for more details.
}

\section{Priors}{

The priors are passed as a named list with elements \code{sigma} and \code{tau}
specifying priors for the corresponding parameters.
Each element of the list should be a vector of length 2, containing values for x1 and x2, x2 > 0, such as:
\itemize{
\item \verb{sigma / max_score ~ normal+(x1, x2)}.
\item \code{tau ~ lognormal(x1, x2)}.
}

NB: For \code{sigma}, usually x1=0 to define a half-normal distribution
since the parameter is constrained to be positive.
}

\section{Default priors}{

\itemize{
\item The default prior for \code{sigma} translates to a width of the predictive distribution to be at most \code{max_score}.
\item The default prior for \code{tau} assumes it could range from less a 1 to 100 (time units).
}
}

\examples{
EczemaModel("Smoothing", max_score = 100)
}
