% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R, R/model_prior.R
\name{validate_prior}
\alias{validate_prior}
\alias{validate_prior.BinRW}
\title{Check the prior of an EczemaModel is correct}
\usage{
validate_prior(model)

\method{validate_prior}{BinRW}(model)
}
\arguments{
\item{model}{Object of class EczemaModel}
}
\value{
NULL if all statements are TRUE, otherwise an error message
}
\description{
Check the prior of an EczemaModel is correct
}
