% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R, R/model_prior.R
\name{validate_prior}
\alias{validate_prior}
\alias{validate_prior.BinRW}
\alias{validate_prior.OrderedRW}
\alias{default_prior.OrderedRW}
\alias{print_prior.OrderedRW}
\title{Check the prior of an EczemaModel is correct}
\usage{
validate_prior(model)

\method{validate_prior}{BinRW}(model)

\method{validate_prior}{OrderedRW}(model)

\method{default_prior}{OrderedRW}(model)

\method{print_prior}{OrderedRW}(model, digits = 2)
}
\arguments{
\item{model}{Object}

\item{...}{Arguments to pass to other methods}
}
\value{
NULL if all statements are TRUE, otherwise an error message
}
\description{
Used internally when constructing an EczemaModel.
}
\examples{
default_prior(EczemaModel("OrderedRW", max_score = 10))
}
