% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R, R/model_prior.R
\name{validate_prior}
\alias{validate_prior}
\alias{validate_prior.BinRW}
\alias{validate_prior.OrderedRW}
\alias{validate_prior.BinMC}
\alias{validate_prior.RW}
\alias{validate_prior.Smoothing}
\alias{validate_prior.AR1}
\alias{validate_prior.MixedAR1}
\alias{validate_prior.MC}
\title{Check the prior of an EczemaModel is correct}
\usage{
validate_prior(model)

\method{validate_prior}{BinRW}(model)

\method{validate_prior}{OrderedRW}(model)

\method{validate_prior}{BinMC}(model)

\method{validate_prior}{RW}(model)

\method{validate_prior}{Smoothing}(model)

\method{validate_prior}{AR1}(model)

\method{validate_prior}{MixedAR1}(model)

\method{validate_prior}{MC}(model)
}
\arguments{
\item{model}{Object}

\item{...}{Arguments to pass to other methods}
}
\value{
NULL if all statements are TRUE, otherwise an error message
}
\description{
Used internally when constructing an EczemaModel.
}
\seealso{
\code{\link[base:stopifnot]{base::stopifnot()}}
}
